feat: Add routes /flights and /flights/hotOffers
Now routes /flights and /flights/hotOffers are working, they give info about specific flights

+++ src/main/kotlin/repository/tables/seatsTables/Seats.kt
@@ -6,7 +6,7 @@ import org.jetbrains.exposed.v1.core.Table
 object Seats : Table("seats") {
     val id = integer("seat_id").autoIncrement()
     val aircraftId = integer("aircraft_id").references(Aircraft.id)
-    val seatClassId = integer("seat_class_id").references(SeatClass.id)
+    val seatClassId = short("seat_class_id").references(SeatClass.id)
     val row = varchar("row", 10)
     val position = varchar("position", 10)
